+++ src/groovy/net/hedtech/restfulapi/config/RepresentationConfig.groovy
@@ -1,5 +1,5 @@
 /* ***************************************************************************
- * Copyright 2013 Ellucian Company L.P. and its affiliates.
+ * Copyright 2013-2015 Ellucian Company L.P. and its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -22,6 +22,7 @@ class RepresentationConfig {
     String contentType
     def jsonArrayPrefix
     def patchSupport
+    String patchAppliesTo
     def marshallers = []
     def extractor