+++ src/groovy/net/hedtech/restfulapi/extractors/json/DefaultJSONExtractor.groovy
@@ -1,5 +1,5 @@
 /* ****************************************************************************
- * Copyright 2013 Ellucian Company L.P. and its affiliates.
+ * Copyright 2013-2015 Ellucian Company L.P. and its affiliates.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -25,6 +25,22 @@ import org.codehaus.groovy.grails.web.json.JSONObject
 class DefaultJSONExtractor implements JSONExtractor {
 
     Map extract( JSONObject content ) {
+        unwrap( content )
+    }
+
+    // We'll build a Map from the JSONObject, as JSONObject
+    // is not serializable and not a direct replacement for Map.
+    //
+    private def unwrap( def content ) {
+        if (content == null) return null
+        if (content == JSONObject.NULL) return null
+        if (content instanceof Map) {
+            Map map = [:]
+            content.entrySet().each { entry ->
+                map[entry.key] = unwrap(entry.value)
+            }
+            return map
+        }
         return content
     }