+++ test/test-restful-api/test/unit/net/hedtech/restfulapi/extractors/json/BasicJSONExtractorSpec.groovy
@@ -32,6 +32,10 @@ import spock.lang.*
 @TestMixin([GrailsUnitTestMixin])
 class BasicJSONExtractorSpec extends Specification {
 
+   def cleanup() {
+       GroovySystem.metaClassRegistry.removeMetaClass BasicJSONExtractor
+   }
+
     def "Test rename paths"() {
         setup:
         BasicJSONExtractor.metaClass.getRenamePaths << {